requirements.txt
fabric~=3.0.0
invoke~=2.0.0
PyYAML~=6.0